Front with products

+++ P-ZPP_1/P-ZPP_1/AppDatabase/Items.cs
@@ -14,13 +14,13 @@ namespace P_ZPP_1.AppDatabase
         public int Id { get; set; }
         public int query_Id { get; set; }
         public string ProductName { get; set; }
-        public decimal Price { get; set; }
+        public float Price { get; set; }
         public float PriceWithShipping { get; set; }
         public int NumberOfPeopleWhoAlsoBought { get; set; }
         public bool ALLEGROsmart { get; set; }
         public string Image { get; set; }
 
-        public Items(int id, int query_Id, string productName, decimal price, float priceWithShipping, int numberOfPeopleWhoAlsoBought, bool aLLEGROsmart, string image)
+        public Items(int id, int query_Id, string productName, float price, float priceWithShipping, int numberOfPeopleWhoAlsoBought, bool aLLEGROsmart, string image)
         {
             Id = id;
             this.query_Id = query_Id;
